refactor: address code review findings for ResourceLoader
- Move ResourceLoader interface from interfaces/ to session/ to eliminate
  circular type dependency (interfaces/ should remain self-contained)
- Consolidate redundant UIProvider passthrough tests into one
- Add test for optional field forwarding (thinkingLevel, extensions)
- Fix import ordering and re-export organization across barrel files
