+++ packages/otter-agent/src/index.ts
@@ -6,7 +6,6 @@ export type {
 	Entry,
 	EntryId,
 	ReadonlySessionManager,
-	ResourceLoader,
 	SessionContext,
 	SessionManagerTemplate,
 	SkillDefinition,
@@ -127,10 +126,10 @@ export {
 export {
 	AgentSession,
 	buildSystemPrompt,
+	createAgentSession,
 	createAgentSessionFromResourceLoader,
 	buildToolSection,
 	convertToLlm,
-	createAgentSession,
 	createCompactionSummaryMessage,
 	createCustomMessage,
 	ModelRegistry,
@@ -143,6 +142,7 @@ export type {
 	CompactionSummaryMessage,
 	CreateAgentSessionOptions,
 	CreateAgentSessionResult,
+	ResourceLoader,
 	CustomMessage,
 } from "./session/index.js";
