+++ packages/otter-agent/src/session/index.ts
@@ -9,6 +9,7 @@ export type {
 	CreateAgentSessionOptions,
 	CreateAgentSessionResult,
 } from "./agent-session.js";
+export type { ResourceLoader } from "./resource-loader.js";
 export { convertToLlm, createCompactionSummaryMessage, createCustomMessage } from "./messages.js";
 export type { CompactionSummaryMessage, CustomMessage } from "./messages.js";
 export { ModelRegistry } from "./model-registry.js";
